# openapi.yaml
openapi: 3.0.3
info:
  title: LifePlace API
  version: 1.0.0
  description: >
    Comprehensive API documentation for LifePlace Event Management Platform.
    Built with Django REST Framework following Domain-Driven Design principles.
    This CRM system manages the complete event lifecycle from booking flows 
    to payment processing, contracts, and client communications.
  contact:
    name: LifePlace API Support
    url: https://api.crm.example.com/support
  license:
    name: Proprietary
    url: https://api.crm.example.com/license
servers:
  - url: https://api.crm.example.com/v1
    description: Production server
  - url: https://staging-api.crm.example.com/v1
    description: Staging server
  - url: http://localhost:8000
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/communications/templates/:
    get:
      tags: [Communications]
      summary: List communication templates
      description: Get all email/SMS templates
      operationId: list_communication_templates
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - name: type
          in: query
          schema:
            type: string
            enum: [email, sms, push]
      responses:
        '200':
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTemplateList'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Communications]
      summary: Create template
      description: Create a new communication template
      operationId: create_template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunicationTemplate'
      responses:
        '201':
          description: Template created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunicationTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/communications/templates/{id}/:
    get:
      tags: [Communications]
      summary: Get template details
      description: Get communication template details
      operationId: get_template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Template retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunicationTemplate'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Communications]
      summary: Update template
      description: Update communication template
      operationId: update_template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommunicationTemplate'
      responses:
        '200':
          description: Template updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommunicationTemplate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Communications]
      summary: Delete template
      description: Delete communication template
      operationId: delete_template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Template deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        default: 1
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string

    # User Management Schemas
...
    CommunicationTemplate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
        type:
          type: string
          enum: [email, sms, push]
        subject:
          type: string
        body:
          type: string
        variables:
          type: array
          items:
            type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - name
        - type
        - body
...
    PaginatedTemplateList:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/CommunicationTemplate'
...